Fixed Infinite loop
Added check for when elevator wanting to go a certain direction, but the
elevator is already past. i.e. request to go to floor 5, and he/she
wants to go down, but elevator is at 1 and is going up. 5 will be added
to queue to go up.
This may create picking up a person and going the wrong way, further
thinking required.